remove unused protocol implementation
